Unify the Coons/Tensor patch-mesh stream loop into a shared DrawPatchMeshUnclipped driver and use the managed MapPointAffine instead of the SKMatrix.MapPoint
