+++ UglyToad.PdfPig.Rendering.Skia/SkiaStreamProcessor.Shading.Coons.cs
@@ -50,177 +50,11 @@ internal partial class SkiaStreamProcessor
 
     private void DrawCoonsMeshUnclipped(CoonsPatchMeshesShading shading, SKMatrix patternTransformMatrix)
     {
-        var currentState = GetCurrentState();
-        int bitsPerCoordinate = shading.BitsPerCoordinate;
-        int bitsPerComponent = shading.BitsPerComponent;
-        int bitsPerFlag = shading.BitsPerFlag;
-        var decode = shading.Decode;
-
-        int numStreamColorComponents = (decode.Length - 4) / 2;
-        double maxCoordRaw = (1L << bitsPerCoordinate) - 1.0;
-        double maxColorRaw = (1L << bitsPerComponent) - 1.0;
-        double xMin = decode[0], xMax = decode[1];
-        double yMin = decode[2], yMax = decode[3];
-
-        // When a Function is present, the colour is non-linear in the bilinear-interpolated
-        // parameter (most visibly: stitched Type-3 functions and Type-2 N=0 step functions).
-        // Per-vertex Gouraud interpolation can't represent these correctly inside a cell —
-        // a cell straddling a step boundary smears the two output colours together.
-        // So for the Function path we draw each patch with a pre-evaluated colour texture
-        // and texture-coordinate mapping, getting per-pixel function output.
-        bool hasFunction = shading.Functions is { Length: > 0 };
-
-        // Per-shading scratch for the no-function (vertex-colour Gouraud) path. Each patch
-        // is submitted via its own indexed DrawVertices call with exact-size vertex arrays,
-        // so memory stays bounded regardless of mesh size.
-        double[]? interpBuffer = null;
-        SKPaint? gouraudPaint = null;
-
-        if (!hasFunction)
-        {
-            interpBuffer = new double[numStreamColorComponents];
-            gouraudPaint = new SKPaint
-            {
-                IsAntialias = shading.AntiAlias,
-                BlendMode = currentState.BlendMode.ToSKBlendMode(),
-                Color = SKColors.White,
-            };
-        }
-
-        // Function path: the per-vertex stream carries a single parametric value, so the colour is
-        // a 1-D function of it. Pre-evaluate that mapping once into a shading-global LUT and reuse
-        // it for every patch texture instead of calling the Function per texel. See BuildPatchTexture.
-        uint[]? patchLut = null;
-        double domainLo = 0d, domainHi = 0d;
-        if (hasFunction && numStreamColorComponents == 1)
-        {
-            domainLo = decode[4];
-            domainHi = decode[5];
-            patchLut = BuildParametricColorLut(shading, currentState, domainLo, domainHi);
-        }
-
-        try
-        {
-            // Patch buffers are alternated between the current and previous patch via a
-            // two-slot pool: the implicit-edge flags (1/2/3) require keeping the previous
-            // patch alive, but at most one previous and one current patch are live at a
-            // time. Pre-allocating both pairs lifts ~12 SKPoint slots + 4 component
-            // arrays out of the per-patch hot loop.
-            var ptsBufA = new SKPoint[12];
-            var ptsBufB = new SKPoint[12];
-            var colorsBufA = new double[4][];
-            var colorsBufB = new double[4][];
-            for (int i = 0; i < 4; i++)
-            {
-                colorsBufA[i] = new double[numStreamColorComponents];
-                colorsBufB[i] = new double[numStreamColorComponents];
-            }
-
-            SKPoint[] points = ptsBufA;
-            double[][] cornerColors = colorsBufA;
-            SKPoint[]? prevPts = null;
-            double[][]? prevColors = null;
-            var bitReader = new GouraudBitReader(shading.Data.Span);
-
-            while (bitReader.HasData)
-            {
-                int flag;
-                try
-                {
-                    flag = (int)(bitReader.ReadBits(bitsPerFlag) & 3);
-                }
-                catch
-                {
-                    break;
-                }
-
-                int newPointCount = flag == 0 ? 12 : 8;
-                int newColorCount = flag == 0 ? 4 : 2;
-
-                if (flag == 0)
-                {
-                    if (!ReadPatchPoints(ref bitReader, bitsPerCoordinate, maxCoordRaw, xMin, xMax, yMin, yMax,
-                            in patternTransformMatrix, points, 0, newPointCount))
-                    {
-                        break;
-                    }
-                    if (!ReadPatchColorsInto(ref bitReader, bitsPerComponent, maxColorRaw, decode, numStreamColorComponents,
-                            cornerColors, 0, newColorCount))
-                    {
-                        break;
-                    }
-                }
-                else
-                {
-                    if (prevPts is null || prevColors is null)
-                    {
-                        // No previous patch — malformed stream; bail out gracefully.
-                        break;
-                    }
-
-                    // Per PDF spec Table 90: the implicit edge of the new patch is the C2 curve of the
-                    // previous patch, the right curve, or the left curve, depending on the flag value.
-                    int p11Idx, p12Idx, p13Idx, p14Idx;        // previous patch boundary points re-used as new patch corners
-                    int newC1ColorIdx, newC2ColorIdx;          // previous patch corner colours that become new patch corner colours
-                    switch (flag)
-                    {
-                        case 1: p11Idx = 3; p12Idx = 4; p13Idx = 5; p14Idx = 6; newC1ColorIdx = 1; newC2ColorIdx = 2;
-                            break;
-                        case 2: p11Idx = 6; p12Idx = 7; p13Idx = 8; p14Idx = 9; newC1ColorIdx = 2; newC2ColorIdx = 3;
-                            break;
-                        case 3: p11Idx = 9; p12Idx = 10; p13Idx = 11; p14Idx = 0; newC1ColorIdx = 3; newC2ColorIdx = 0;
-                            break;
-                        default:
-                            return;
-                    }
-
-                    points[0] = prevPts[p11Idx];
-                    points[1] = prevPts[p12Idx];
-                    points[2] = prevPts[p13Idx];
-                    points[3] = prevPts[p14Idx];
-
-                    // Copy component values from prev's slot into current's slot — the
-                    // destination array is already owned by `cornerColors`, so we don't
-                    // reassign the slot reference (that would alias prev's buffer and the
-                    // next patch would overwrite both).
-                    Array.Copy(prevColors[newC1ColorIdx], cornerColors[0], numStreamColorComponents);
-                    Array.Copy(prevColors[newC2ColorIdx], cornerColors[1], numStreamColorComponents);
-
-                    if (!ReadPatchPoints(ref bitReader, bitsPerCoordinate, maxCoordRaw, xMin, xMax, yMin, yMax,
-                            in patternTransformMatrix, points, 4, newPointCount))
-                    {
-                        break;
-                    }
-                    if (!ReadPatchColorsInto(ref bitReader, bitsPerComponent, maxColorRaw, decode, numStreamColorComponents,
-                            cornerColors, 2, newColorCount))
-                    {
-                        break;
-                    }
-                }
-
-                bitReader.AlignToByte();
-
-                if (hasFunction)
-                {
-                    DrawCoonsPatchTextured(shading, currentState, points, cornerColors, patchLut, domainLo, domainHi);
-                }
-                else
-                {
-                    TessellateAndDrawCoonsPatch(shading, currentState, points, cornerColors,
-                        interpBuffer!, gouraudPaint!);
-                }
-
-                prevPts = points;
-                prevColors = cornerColors;
-                // Alternate the active buffer so prev stays valid while we fill current.
-                points = ReferenceEquals(points, ptsBufA) ? ptsBufB : ptsBufA;
-                cornerColors = ReferenceEquals(cornerColors, colorsBufA) ? colorsBufB : colorsBufA;
-            }
-        }
-        finally
-        {
-            gouraudPaint?.Dispose();
-        }
+        // The stream layout and edge-continuation rules are shared with Type 7 — see
+        // DrawPatchMeshUnclipped for the loop; only the per-patch tessellation is Coons-specific.
+        DrawPatchMeshUnclipped(shading, in patternTransformMatrix, shading.Data.Span,
+            shading.BitsPerCoordinate, shading.BitsPerComponent, shading.BitsPerFlag,
+            shading.Decode, isTensor: false);
     }
 
     /// <summary>
