+++ UglyToad.PdfPig.Rendering.Skia/SkiaStreamProcessor.Shading.Tensor.cs
@@ -56,155 +56,11 @@ internal partial class SkiaStreamProcessor
     private void DrawTensorMeshUnclipped(TensorProductPatchMeshesShading shading,
       SKMatrix patternTransformMatrix)
     {
-        var currentState = GetCurrentState();
-        int bitsPerCoordinate = shading.BitsPerCoordinate;
-        int bitsPerComponent = shading.BitsPerComponent;
-        int bitsPerFlag = shading.BitsPerFlag;
-        var decode = shading.Decode;
-
-        int numStreamColorComponents = (decode.Length - 4) / 2;
-        double maxCoordRaw = (1L << bitsPerCoordinate) - 1.0;
-        double maxColorRaw = (1L << bitsPerComponent) - 1.0;
-        double xMin = decode[0], xMax = decode[1];
-        double yMin = decode[2], yMax = decode[3];
-
-        // See RenderCoonsPatchShading for why the function path uses textured drawing.
-        bool hasFunction = shading.Functions is { Length: > 0 };
-
-        // Per-shading scratch for the no-function (vertex-colour Gouraud) path. See
-        // RenderCoonsPatchShading for the per-patch indexed-DrawVertices rationale.
-        double[]? interpBuffer = null;
-        SKPaint? gouraudPaint = null;
-
-        if (!hasFunction)
-        {
-            interpBuffer = new double[numStreamColorComponents];
-            gouraudPaint = new SKPaint
-            {
-                IsAntialias = shading.AntiAlias,
-                BlendMode = currentState.BlendMode.ToSKBlendMode(),
-                Color = SKColors.White,
-            };
-        }
-
-        // See RenderCoonsPatchShading: pre-evaluate the 1-D parametric colour mapping once into a
-        // shading-global LUT instead of calling the Function per texel of every patch texture.
-        uint[]? patchLut = null;
-        double domainLo = 0d, domainHi = 0d;
-        if (hasFunction && numStreamColorComponents == 1)
-        {
-            domainLo = decode[4];
-            domainHi = decode[5];
-            patchLut = BuildParametricColorLut(shading, currentState, domainLo, domainHi);
-        }
-
-        try
-        {
-            // See RenderCoonsPatchShading for the two-slot ring-buffer rationale.
-            var ptsBufA = new SKPoint[16];
-            var ptsBufB = new SKPoint[16];
-            var colorsBufA = new double[4][];
-            var colorsBufB = new double[4][];
-            for (int i = 0; i < 4; i++)
-            {
-                colorsBufA[i] = new double[numStreamColorComponents];
-                colorsBufB[i] = new double[numStreamColorComponents];
-            }
-
-            SKPoint[] points = ptsBufA;
-            double[][] cornerColors = colorsBufA;
-            SKPoint[]? prevPts = null;
-            double[][]? prevColors = null;
-            var bitReader = new GouraudBitReader(shading.Data.Span);
-
-            while (bitReader.HasData)
-            {
-                int flag;
-                try
-                {
-                    flag = (int)(bitReader.ReadBits(bitsPerFlag) & 3);
-                }
-                catch
-                {
-                    break;
-                }
-
-                int newPointCount = flag == 0 ? 16 : 12;
-                int newColorCount = flag == 0 ? 4 : 2;
-
-                if (flag == 0)
-                {
-                    if (!ReadPatchPoints(ref bitReader, bitsPerCoordinate, maxCoordRaw, xMin, xMax, yMin, yMax,
-                            in patternTransformMatrix, points, 0, newPointCount))
-                    {
-                        break;
-                    }
-                    if (!ReadPatchColorsInto(ref bitReader, bitsPerComponent, maxColorRaw, decode, numStreamColorComponents,
-                            cornerColors, 0, newColorCount))
-                    {
-                        break;
-                    }
-                }
-                else
-                {
-                    if (prevPts is null || prevColors is null)
-                    {
-                        break;
-                    }
-
-                    // For Type 7, only the four boundary corners (indices 0, 3, 6, 9 in the 16-point sequence)
-                    // are reused; the four interior control points (indices 12-15) are always new.
-                    int p11Idx, p12Idx, p13Idx, p14Idx;
-                    int newC1ColorIdx, newC2ColorIdx;
-                    switch (flag)
-                    {
-                        case 1: p11Idx = 3; p12Idx = 4; p13Idx = 5; p14Idx = 6; newC1ColorIdx = 1; newC2ColorIdx = 2; break;
-                        case 2: p11Idx = 6; p12Idx = 7; p13Idx = 8; p14Idx = 9; newC1ColorIdx = 2; newC2ColorIdx = 3; break;
-                        case 3: p11Idx = 9; p12Idx = 10; p13Idx = 11; p14Idx = 0; newC1ColorIdx = 3; newC2ColorIdx = 0; break;
-                        default: return;
-                    }
-
-                    points[0] = prevPts[p11Idx];
-                    points[1] = prevPts[p12Idx];
-                    points[2] = prevPts[p13Idx];
-                    points[3] = prevPts[p14Idx];
-                    Array.Copy(prevColors[newC1ColorIdx], cornerColors[0], numStreamColorComponents);
-                    Array.Copy(prevColors[newC2ColorIdx], cornerColors[1], numStreamColorComponents);
-
-                    if (!ReadPatchPoints(ref bitReader, bitsPerCoordinate, maxCoordRaw, xMin, xMax, yMin, yMax,
-                            in patternTransformMatrix, points, 4, newPointCount))
-                    {
-                        break;
-                    }
-                    if (!ReadPatchColorsInto(ref bitReader, bitsPerComponent, maxColorRaw, decode, numStreamColorComponents,
-                            cornerColors, 2, newColorCount))
-                    {
-                        break;
-                    }
-                }
-
-                bitReader.AlignToByte();
-
-                if (hasFunction)
-                {
-                    DrawTensorPatchTextured(shading, currentState, points, cornerColors, patchLut, domainLo, domainHi);
-                }
-                else
-                {
-                    TessellateAndDrawTensorPatch(shading, currentState, points, cornerColors,
-                        interpBuffer!, gouraudPaint!);
-                }
-
-                prevPts = points;
-                prevColors = cornerColors;
-                points = ReferenceEquals(points, ptsBufA) ? ptsBufB : ptsBufA;
-                cornerColors = ReferenceEquals(cornerColors, colorsBufA) ? colorsBufB : colorsBufA;
-            }
-        }
-        finally
-        {
-            gouraudPaint?.Dispose();
-        }
+        // The stream layout and edge-continuation rules are shared with Type 6 — see
+        // DrawPatchMeshUnclipped for the loop; only the per-patch tessellation is Tensor-specific.
+        DrawPatchMeshUnclipped(shading, in patternTransformMatrix, shading.Data.Span,
+            shading.BitsPerCoordinate, shading.BitsPerComponent, shading.BitsPerFlag,
+            shading.Decode, isTensor: true);
     }
 
     /// <summary>
